ui_v5/requirements.txt
nicegui[pydantic]
httpx
python-dotenv
sqlalchemy
pymysql
cryptography
